entreVoisins-master/requirements.txt
Django==3.0.7
djangorestframework==3.6.2

